--- src/HANReader.Core/IFrameReader.cs
@@ -1,10 +1,0 @@
-using System.Buffers;
-using HANReader.Core.Models;
-
-namespace HANReader.Core
-{
-    public interface IFrameReader
-    {
-        ReadStatus ReadFrame(ref ReadOnlySequence<byte> buffer, out Frame frame);
-    }
-}
